Add user table insert, update, and delete handlers
Introduces new modules for user table INSERT, UPDATE, and DELETE operations with data isolation, system column management, and RocksDB integration. Refactors table_cache and catalog_store to support RocksDB and kalamdb-sql, updates service and test code for new architecture, and re-exports new handlers in tables/mod.rs.

+++ backend/crates/kalamdb-core/src/services/namespace_service.rs
@@ -105,7 +105,7 @@ impl NamespaceService {
                     created_at: chrono::DateTime::from_timestamp(ns.created_at, 0)
                         .ok_or_else(|| KalamDbError::IoError("Invalid timestamp".to_string()))?,
                     options,
-                    table_count: ns.table_count as usize,
+                    table_count: ns.table_count as u32,
                 }))
             }
             None => Ok(None),
